voting_service/requirements.txt
Flask==2.0.3
Flask-RESTful==0.3.9
psycopg2-binary==2.9.3
Werkzeug==2.0.3
Flask-Cors==3.0.10
